+++ src/widgets/mod.rs
@@ -4,6 +4,6 @@ pub mod dock_area;
 /// Trait for tab-viewing types.
 pub mod tab_viewer;
 
-pub use dock_area::ids::{dragged_tab, tab_widget_id};
+pub use dock_area::ids::{drag_hover_node, dragged_tab, tab_widget_id};
 pub use dock_area::{AllowedSplits, DockArea};
 pub use tab_viewer::TabViewer;